Moved MarkovSequences
