--- docs/api_docs_advanced/solution.md
@@ -1,2 +1,0 @@
-# Solution objects
-::: odefilter.solution
