+++ docs/api_docs_advanced/solutions.md
@@ -1,0 +1,2 @@
+# Solution objects
+::: odefilter.solutions
